:fire: Merge new variants
