+++ src/graphql/variants/actions.ts
@@ -4,7 +4,6 @@ import {
   IQueryVariable,
 } from '@ferlab/ui/core/graphql/types';
 import { computeSearchAfter, hydrateResults } from '@ferlab/ui/core/graphql/utils';
-import { IVariantEntity as IVariantEntityFerlab } from '@ferlab/ui/core/pages/EntityPage/type';
 import { INDEXES } from 'graphql/constants';
 
 import useLazyResultQuery from 'hooks/graphql/useLazyResultQuery';
@@ -35,7 +34,7 @@ interface IUseVariantEntityProps {
 
 interface IUseVariantEntityReturn {
   loading: boolean;
-  data?: IVariantEntityFerlab;
+  data?: IVariantEntity;
 }
 
 export const useVariantEntity = ({
@@ -51,7 +50,7 @@ export const useVariantEntity = ({
     variables: { sqon },
   });
 
-  const data = result?.variants?.hits?.edges[0].node;
+  const data = result?.variants?.hits?.edges[0]?.node;
 
   return {
     loading,
