+++ src/views/VariantEntity/FerlabComponent/EntityPublicCohortTable.tsx
@@ -1,6 +1,6 @@
 import { EntityTable, IEntityTable } from '@ferlab/ui/core/pages/EntityPage';
 import { isExternalCohortsTableEmpty } from '@ferlab/ui/core/pages/EntityPage/utils/frequencies';
-import { IExternalFrequenciesEntity } from 'graphql/variantsv2/models';
+import { IExternalFrequenciesEntity } from '../../../graphql/variants/models';
 
 import { makeRowFromFrequencies } from './EntityPublicCohortTable.utils';
