--- src/views/VariantEntityV2/index.module.scss
@@ -1,32 +1,0 @@
-@import 'src/style/themes/include/theme.template';
-
-.predictionLabel {
-  margin-right: 4px;
-}
-
-.canonicalIcon {
-  color: $canonical-icon-color;
-}
-
-.dotted {
-  border-bottom: 1px dotted;
-}
-
-.variantTag {
-  color: $variant-tag-color;
-  background-color: $variant-tag-bg-color;
-}
-
-.titleIcon {
-  font-size: 24px;
-}
-
-.geneExternalLink {
-  &:not(&:first-child) {
-    &::before {
-      content: ',';
-      margin-right: 4px;
-      color: $body-1;
-    }
-  }
-}
