+++ src/views/Variants/components/ConsequencesCell/index.tsx
@@ -2,15 +2,13 @@ import ExternalLink from '@ferlab/ui/core/components/ExternalLink';
 import { IArrangerEdge } from '@ferlab/ui/core/graphql/types';
 import StackLayout from '@ferlab/ui/core/layout/StackLayout';
 import { removeUnderscoreAndCapitalize } from '@ferlab/ui/core/utils/stringUtils';
-import { IConsequenceEntity, Impact } from 'graphql/variants/models';
+import { IConsequenceEntity, Impact } from '../../../../graphql/variants/models';
 
 import HighBadgeIcon from 'components/Icons/VariantBadgeIcons/HighBadgeIcon';
 import LowBadgeIcon from 'components/Icons/VariantBadgeIcons/LowBadgeIcon';
 import ModerateBadgeIcon from 'components/Icons/VariantBadgeIcons/ModerateBadgeIcon';
 import ModifierBadgeIcon from 'components/Icons/VariantBadgeIcons/ModifierBadgeIcon';
 import { toKebabCase } from 'utils/helper';
-
-import { generateConsequencesDataLines } from './consequences';
 
 import style from './index.module.scss';
 
@@ -25,40 +23,37 @@ const pickImpactBadge = (impact: Impact) => impactToColorClassName[impact];
 
 const ConsequencesCell = ({
   consequences,
+  symbol,
 }: {
   consequences: IArrangerEdge<IConsequenceEntity>[];
+  symbol: string;
 }) => {
-  const lines = generateConsequencesDataLines(consequences);
+  const pickedEdge = consequences.find((c) => c.node.picked);
+  if (!pickedEdge) {
+    //must never happen
+    return null;
+  }
+
+  const picked = pickedEdge?.node;
+  if (!picked?.consequence) {
+    return null;
+  }
+
   return (
-    <>
-      {lines.map((consequence, index) => {
-        /* Note: index can be used as key since the list is readonly */
-        const node = consequence.node;
-
-        if (node.consequences) {
-          return (
-            <StackLayout className={style.stackLayout} center key={index}>
-              {pickImpactBadge(node.vep_impact)}
-              <span key={index} className={style.detail}>
-                {removeUnderscoreAndCapitalize(node.consequences[0])}
-              </span>
-              {node.symbol && (
-                <span key={toKebabCase(node.symbol)} className={style.symbol}>
-                  <ExternalLink
-                    href={`https://useast.ensembl.org/Homo_sapiens/Gene/Summary?g=${node.symbol}`}
-                  >
-                    {node.symbol}
-                  </ExternalLink>
-                </span>
-              )}
-              {node.hgvsp && <span>{node.hgvsp.split(':')[1]}</span>}
-            </StackLayout>
-          );
-        }
-
-        return null;
-      })}
-    </>
+    <StackLayout className={style.stackLayout} center key={'1'}>
+      {pickImpactBadge(picked.vep_impact)}
+      <span key="detail" className={style.detail}>
+        {removeUnderscoreAndCapitalize(picked.consequence[0])}
+      </span>
+      {symbol && (
+        <span key={toKebabCase(symbol)} className={style.symbol}>
+          <ExternalLink href={`https://useast.ensembl.org/Homo_sapiens/Gene/Summary?g=${symbol}`}>
+            {symbol}
+          </ExternalLink>
+        </span>
+      )}
+      {picked.hgvsp && <span>{picked.hgvsp.split(':')[1]}</span>}
+    </StackLayout>
   );
 };
