+++ src/views/Variants/index.module.scss
@@ -28,9 +28,5 @@
   .scrollContent {
     width: 100%;
     padding: $default-page-content-padding;
-
-    .title {
-      margin-bottom: 16px;
-    }
   }
 }
