--- src/views/VariantsV2/components/ConsequencesCell/index.module.scss
@@ -1,40 +1,0 @@
-@import 'src/style/themes/include/colors';
-
-.stackLayout {
-  min-width: 375px;
-}
-
-.consequenceUl {
-  list-style: none;
-  padding: 0;
-}
-
-.highImpact {
-  fill: $red-5;
-}
-
-.lowImpact {
-  fill: $green-6;
-}
-
-.moderateImpact {
-  fill: $orange-5;
-}
-
-.modifierImpact {
-  fill: $gray-6;
-}
-
-.detail {
-  padding-right: 12px;
-}
-
-.bullet {
-  margin-right: 8px;
-  min-width: 10px;
-  min-height: 10px;
-}
-
-.symbol {
-  padding-right: 8px;
-}
