--- src/views/VariantsV2/components/GeneUploadIds/index.module.scss
@@ -1,11 +1,0 @@
-.geneUploadIdsPopover {
-    max-width: 400px;
-
-    :global(.ant-descriptions-item) {
-        padding-bottom: 0;
-    }
-
-    :global(.ant-descriptions-item-label) {
-        font-weight: 500;
-    }
-}
