--- src/views/VariantsV2/components/PageContent/VariantsTable/index.module.scss
@@ -1,16 +1,0 @@
-.variantTabWrapper {
-  display: flex;
-
-  :global(.ant-pagination) {
-    margin-bottom: 0;
-  }
-}
-
-.variantTableCellElipsis {
-  max-width: 200px;
-  width: 200px;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  overflow-wrap: unset;
-  white-space: nowrap;
-}
