--- src/views/VariantsV2/components/PageContent/index.module.scss
@@ -1,36 +1,0 @@
-@import 'src/style/themes/include/colors';
-
-.pageHeader {
-  display: flex;
-  align-items: center;
-
-  .pageHeaderTitle {
-    display: flex;
-    margin: 0;
-    font-size: 20px;
-    line-height: 28px;
-    margin-right: 1.2rem;
-  }
-
-  .dataReleaseTag {
-    color: $cyan-7;
-    display: flex;
-    align-items: center;
-    font-size: 12px;
-    height: 20px;
-
-    &:hover {
-      cursor: pointer;
-    }
-
-    span {
-      margin-right: 0.5rem;
-    }
-  }
-}
-
-.variantsPageContent {
-  width: 100%;
-  display: flex;
-  position: relative;
-}
