--- src/views/VariantsV2/components/VariantGeneSearch/OptionItem/index.module.scss
@@ -1,15 +1,0 @@
-@import 'src/style/themes/include/colors';
-
-.searchLabel {
-  &.variants {
-    background-color: $purple-2;
-  }
-
-  &.genes {
-    background-color: $geekblue-2;
-  }
-}
-
-.variantSearchLocus {
-  font-weight: 500;
-}
