--- src/views/VariantsV2/index.module.scss
@@ -1,32 +1,0 @@
-@import 'src/style/themes/include/layout';
-
-.variantsLayout {
-  display: flex;
-  height: calc(100vh - #{$main-header-height});
-
-  .filterLoader {
-    padding: 24px;
-    width: 100%;
-  }
-
-  .sideMenu {
-    li[role='menuitem'] {
-      color: white;
-
-      span[class$='-title-content'] {
-        margin-left: 0 !important;
-      }
-    }
-  }
-
-  span[role='img'].sideMenuIcon {
-    font-size: 20px !important;
-    display: flex;
-    justify-content: center;
-  }
-
-  .scrollContent {
-    width: 100%;
-    padding: $default-page-content-padding;
-  }
-}
